Changes on 09/08

+++ src/main/java/com/example/E_commerce/Customer/service/CustomerService.java
@@ -3,6 +3,7 @@ package com.example.E_commerce.Customer.service;
 import com.example.E_commerce.Customer.assembler.CustomerAssembler;
 import com.example.E_commerce.Customer.dto.CustomerRequestDTO;
 import com.example.E_commerce.Customer.dto.CustomerResponseDTO;
+import com.example.E_commerce.Exception.CustomerNotFoundException;
 import com.example.E_commerce.Persistance.utils.CustomerMapper;
 import com.example.E_commerce.Persistance.model.Customer;
 import com.example.E_commerce.Persistance.repository.CustomerRepository;
@@ -15,18 +16,20 @@ import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
+import static com.example.E_commerce.Constants.CommonConstants.C_NOTFOUND;
 
 @Service
 public class CustomerService {
 
-    @Autowired
-    private CustomerRepository customerRepository;
+    private final CustomerRepository customerRepository;
+    private final CustomerMapper customerMapper;
+    private final CustomerAssembler customerAssembler;
 
-    @Autowired
-    private CustomerMapper customerMapper;
-
-    @Autowired
-    private CustomerAssembler customerAssembler;
+    public CustomerService(CustomerRepository customerRepository, CustomerMapper customerMapper, CustomerAssembler customerAssembler) {
+        this.customerRepository = customerRepository;
+        this.customerMapper = customerMapper;
+        this.customerAssembler = customerAssembler;
+    }
 
     public Page<CustomerResponseDTO> getAllCustomers(int page, int size, String sortBy, String sortOrder) {
         Sort sort = Sort.by(Sort.Direction.fromString(sortOrder), sortBy);
@@ -36,11 +39,11 @@ public class CustomerService {
 
     public CustomerResponseDTO getCustomerById(long id) {
         Customer customer = customerRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Customer not found"));
+                .orElseThrow(() -> new CustomerNotFoundException(C_NOTFOUND));
         return customerMapper.toResponseDTO(customer);
     }
 
-    public CustomerResponseDTO saveCustomer(CustomerRequestDTO dto) {
+    public CustomerResponseDTO saveCustomer(CustomerRequestDTO dto) { //Validation need to be done here
         Customer customer = CustomerMapper.toEntity(dto);
         Customer saved = customerRepository.save(customer);
         return customerMapper.toResponseDTO(saved);
@@ -49,8 +52,7 @@ public class CustomerService {
     public CustomerResponseDTO updateCustomer(Long id, CustomerRequestDTO dto)
     {
             Optional<Customer> customerOptional = customerRepository.findById(id);
-            if(customerOptional.isPresent())
-            {
+            if(customerOptional.isPresent()) {
                 Customer customer = customerOptional.get();
                 customer.setCName(dto.getCName());
                 customer.setCEmail(dto.getCEmail());
@@ -60,11 +62,14 @@ public class CustomerService {
                 return customerMapper.toResponseDTO(customerRepository.save(customer));
             }
             else {
-                throw new RuntimeException("Customer with ID " + id + " not found");
+                throw new CustomerNotFoundException(C_NOTFOUND);
             }
     }
 
     public void deleteCustomer(Long id) {
+        if(!customerRepository.existsById(id)) {
+            throw new CustomerNotFoundException(C_NOTFOUND);
+        }
         customerRepository.deleteById(id);
     }
